+++ R/grid-lookup.R
@@ -9,7 +9,7 @@
 #'
 #' @param gridObj (\code{\link{trigrid}} or \code{\link{hexagrid}}) An icoshedral grid.
 #' 
-#' @param data (\code{matrix}, \code{data.frame} or \code{Spatial}) The queried data.
+#' @param data (\code{matrix}, \code{data.frame}, \code{sf} or \code{Spatial}) The queried data.
 #' @param out (\code{character}) What shall be the output class? Can be either \code{\link{facelayer}} or \code{logical} (default.)
 #'
 #' @param ... Arguments passed to the class specific methods
@@ -245,6 +245,19 @@ setMethod(
 	}
 )
 
+
+setMethod(
+	"OccupiedFaces",
+	signature=c("trigrid", "sfc"),
+	definition=function(gridObj, data){
+		temp<-methods::as(data,"Spatial")
+
+		
+		# this works for spatialpolygons and spatialpolygonsdataframes
+		fl <- OccupiedFaces(gridObj, temp)
+		return(fl)
+	}
+)
 
 #for sf: fall to SpatialPolygonsDataFrame
 setMethod(